fix: Adding / Removing favorites is very slow #3
